requirements.txt
numpy
pandas
kafka-python
matplotlib
tqdm
pytest